[hir2mpl]Internal commit msg:
    cg_bugfix_add_support_for_aarch64_aapcs_B.6
    int128-va-fix
    [mplfe] codecheck fix
    [asan attr add key]
    [enhance c]bugfix funcPtr have index and no index attr at the same time
    [ICSL] Clean Code 65
    [mplfe] enable lto fm
    [hir2mpl] support partOfExplicitCast
    [hir2mpl]support flexible arrays enhance c attribute
    open llvm macro in gn make hir2mpl
    [ICSL] Clean Code 63
    [hir2mpl][enhancec] fix returnassertle in func multibranch returns
    [ICSL] Clean Code 62
    [hir2mpl] fix mulalign offset condition
    [ICSL] Clean Code 60
    [ICSL] Clean Code 61
    [hir2mpl]fix lto no delete origin func declare
    [ 6 astParser add allocator, contain 4]
    [hir2mpl] add addrofconstarray const attr
    [ICSL] Clean Code 58
    [hir2mpl] fix typealign set error
    hir2mpl_bugfix_add_some_builtin_cvt_func_for_FORTIFY_SOURCE_2
    [hir2mpl] fix noreturnfunc mistake
    [hir2mpl] fix fieldname construct mistake
    [ICSL] Clean Code 57
[mapleall]Internal commit msg:
    [litepgo] spec pgo bugfix
    [cg] ra loop split in nestedloop bugfix
    dup ret phase
    [maplecg] Add CselToCsincRemoveMovPattern in PeepHole
    cg_bugfix_add_support_for_aarch64_aapcs_B.6
    [cg] fast path opt
    do not get phi node for volatile symbols
    [mapleall] Optimize slp compilation time by skipping huge store sequence
    cg_bugfix_disable_implicit_trunc_in_ExtValidBitPattern
    [ICSL] Clean Code 66
    [driver]delete lto option
    [codecheck me]
    [mplme] fix bug of band opt
    [cg] fix f128 register alloc
    Fix codegen for int128 bitfields
    alarm fixes
    [cg issue217 add A64ConstProp::MovLslConstToMov /* mov w1, #1 lsl w1, w1, #6 ===> mov w1, #xx */]
    validbit opt enhance ubfiz and pattern
    [mapleall] cleancode2  63
    [driver]Do not use Linux commands directly
    cg_bugfix_fix_ra_insn_count&peep&cfgo
    [mapleall] Bugfix for self-recursive always inline func when rai is
    [mapleall] cleancode2  62
    [cg] mpltime ra codestyle
    [mplcg] no strldropt for ldr stImmOperand with wrong alignment
    [cg] add time tool and reduce ra time
    [driver]clean code
    [mplme] fix bug in computing known bits
    [util] -time-phases update
    [mplcg] Modify immediate validrange of tbz 32-bits
    [cg] RA loop around split bugfix
    [driver]enable specs by gcc
    [ICSL] Clean Code 65
    [mapleall] cleancode2  61
    set respectAlwaysInline on by default
    [mapleall] Fix slp cvt bug
    cg_bugfix_fix_ElimSpecificExtensionPattern_condition
    [litepgo] Modify firstbb of func when firstbb in cold section
    [mapleall] fix vrp deal with cvt
    fix implicit cvt like "add i32 (u1, i32)"
    [mplcg] clear insn base attr
    [mapleall] cleancode2  60
    [mplfe] enable lto fm
    [cg] lower bit field read bugfix
    [driver] -DNDEBUG
    do not prop reg replaced union var
    [mplme] cfgo cmp fold bugfix
    [util] bugfix when using -time-phases
    [driver] change option which begin with "--" to "-"
    [issue162] enhance ReplaceCmpToCmn in Constprop
    [codecheck 627 new]
    [me] get field type align
    code style
    [mplcg] Split new invalid insn in framefinalize
    replace small union var with register
    [mplme] fix band optimize bug
    [mplme] cmp fold bugfix
    [cg]clean code
    [cg] agg param copy offset bugfix`
    Legalize numeric types
    remove redundant check fatal
    [cg] register limit clean code
    [asan fix vrp-min]
    [ICSL] Clean Code 64
    [alias] pass maydef/use to cg
    [mapleall] cleancode2  59
    [alias] prop next level nads for escaped agg
    add comment
    [cg] fix retval in agg
    [ICSL] Clean Code 63
    [driver] spec testsuite
    [cg] enable register limit
    [mapleall] cleancode2  58
    [cg]Align bb by freq
    [mapleall] deal with iaddrof in scalar analysis
    [ICSL] Clean Code 62
    [mplcg] cgtargeprop not working when -fpie/-fPIE is enabled
    [mplme] bugfix of fold cmp
    [mpl2mpl][bugfix] Skip not-supported constant folding for f128
    [cg]clean code
    [codecheck mplintVal kconst]
    [mapleall] bugfix vrp
    [ICSL] Clean Code 60
    [ICSL] Clean Code 59
    [codecheck me cg-format]
    [ICSL] Clean Code 61
    [alias] update alias attr correctly in ddaa
    [cg]Insert nop to align loop
    [mplme] add combine expr phase
    [cg] ra spill def-use opnd in large func bugfix
    [simplify] adjust interfaces
    [hir2mpl]fix lto no delete origin func declare
    [mapleall] cleancode2  57
    cg_bugfix_always_sweep_static_symbol
    duplicate code and bug fix
    [mapleall] bugfix objsize
    [mplcg] Modify valid range of tbnz 32-bits immediate
    [cg]Bugfix for ico
    [mplcg] fix bug of checking validDefInsns in PregPropOpt
    [mplme] fix infinite recursion in  stack var addr check
    [alias] apply nads mayuse sets to membarriers
    [mapleall] cleancode2  56
    [me] struct align
    [codecheck orig_symbol.h int32 + need long]
    [mapleall] clean code
    [mapleall] Fix castopt implicit cast bug
    [mapleall] Add mpl2mpl phase CallTargetReplace to replace mangled func that must
    [cgoption] fix CanDoNoPlt needs to check mirFunc cannot be nullptr
    [mplcg] Enhance AddSubMergeLdStPattern in peep
    [mapleall] cleancode2  55
    [ipa] side effect reinforcement
    [codecheck optimizeCFG litepgo mpl_int_val]
    [cg] copy prop cannot prop destreg which was used as use-def
    [driver]Wopt
    [mplcg] Bugfix for non-calleesaved reg used in pgogen
    [mplcg] fix bug of implicit-cvt  in ExtendShiftPattern
    [option] add unified function for checking no-plt enable conditions
    [codecheck]
    [codecheck c17]
    [me_ir.h add private assignCSTR]
    Support int128 in AArch64 Codegen  
    fix insert after switch
    [mapleall] bugfix ivopt
    fix may throw exception expr out-loop bug
    cg_bugfix_adds_imm_valid
    [mapleall] cleancode2  54
    [ICSL] Clean Code 57
    [ICSL] Clean Code 56
    [litepgo] Bugfix for pgouse hotcoldsplit with infinite loop
    [maplecg] add vra register use infomation for propInmem during cgtargetprop phase

+++ src/mapleall/maple_me/src/seqvec.cpp
@@ -536,7 +536,7 @@ bool SeqVectorize::CanSeqVec(const IassignNode *s1, const IassignNode *s2, bool 
 }
 
 static uint32 PreviousPowerOfTwo(unsigned int x) {
-  return 1U << ((sizeof(x) * 8 - 1) - static_cast<uint64>(__builtin_clz(x)));
+  return 1U << ((sizeof(x) * 8 - 1) - static_cast<uint>(__builtin_clz(x)));
 }
 
 void SeqVectorize::MergeIassigns(MapleVector<IassignNode *> &cands) {
